+++ booby/fields.py
@@ -32,7 +32,14 @@ The example below shows the most common fields and builtin validations::
         is_active = Boolean(default=False)
 """
 
+import six
 import collections
+
+try:
+    # Python 3 compatiblity
+    from urlparse import urlparse
+except ImportError:
+    from urllib.parse import urlparse
 
 from booby import (
     validators as builtin_validators,
@@ -67,17 +74,20 @@ class Field(object):
         self.options = kwargs
 
         self.default = kwargs.get('default')
+        self.description = kwargs.get('description', '')
+        self.required = kwargs.get('required', False)
+        self.choices = kwargs.get('choices', [])
+
+        assert isinstance(self.choices, list)
 
         # Setup field validators
         self.validators = []
 
-        if kwargs.get('required'):
+        if self.required:
             self.validators.append(builtin_validators.Required())
 
-        choices = kwargs.get('choices')
-
-        if choices:
-            self.validators.append(builtin_validators.In(choices))
+        if self.choices:
+            self.validators.append(builtin_validators.In(self.choices))
 
         self.validators.extend(validators)
 
@@ -100,6 +110,8 @@ class Field(object):
         return self
 
     def __set__(self, instance, value):
+        if not value:
+            value = self.default
         instance._data[self] = value
 
     def _default(self, model):
@@ -133,12 +145,24 @@ class Field(object):
 
         return value
 
+    @property
+    def field_type(self):
+        """
+        :return: Python native type: int, str, float...
+        :rtype: object
+        """
+        raise NotImplemented("This property must be implemented by the subclass")
+    
 
 class String(Field):
     """:class:`Field` subclass with builtin `string` validation."""
 
     def __init__(self, *args, **kwargs):
         super(String, self).__init__(builtin_validators.String(), *args, **kwargs)
+
+    @property
+    def field_type(self):
+        return six.string_types
 
 
 class Integer(Field):
@@ -147,6 +171,10 @@ class Integer(Field):
     def __init__(self, *args, **kwargs):
         super(Integer, self).__init__(builtin_validators.Integer(), *args, **kwargs)
 
+    @property
+    def field_type(self):
+        return six.integer_types
+
 
 class Float(Field):
     """:class:`Field` subclass with builtin `float` validation."""
@@ -154,6 +182,10 @@ class Float(Field):
     def __init__(self, *args, **kwargs):
         super(Float, self).__init__(builtin_validators.Float(), *args, **kwargs)
 
+    @property
+    def field_type(self):
+        return float
+    
 
 class Boolean(Field):
     """:class:`Field` subclass with builtin `bool` validation."""
@@ -161,6 +193,10 @@ class Boolean(Field):
     def __init__(self, *args, **kwargs):
         super(Boolean, self).__init__(builtin_validators.Boolean(), *args, **kwargs)
 
+    @property
+    def field_type(self):
+        return bool
+    
 
 class Embedded(Field):
     """:class:`Field` subclass with builtin embedded :class:`models.Model`
@@ -183,11 +219,39 @@ class Embedded(Field):
         super(Embedded, self).__set__(instance, value)
 
 
-class Email(Field):
+class Email(String):
     """:class:`Field` subclass with builtin `email` validation."""
 
     def __init__(self, *args, **kwargs):
         super(Email, self).__init__(builtin_validators.Email(), *args, **kwargs)
+
+
+class URL(String):
+    """:class:`Field` subclass with builtin `URL` validation."""
+
+    def __init__(self, *args, **kwargs):
+        super(URL, self).__init__(builtin_validators.URL(), *args, **kwargs)
+
+
+class IP(String):
+    """:class:`Field` subclass with builtin `ip` validation."""
+
+    def __init__(self, *args, **kwargs):
+        super(IP, self).__init__(builtin_validators.IP(), *args, **kwargs)
+
+
+class URI(String):
+    """:class:`Field` subclass with builtin `URI` validation."""
+    
+    def __init__(self, *args, **kwargs):
+        super(URI, self).__init__(builtin_validators.URI(), *args, **kwargs)
+
+
+class Raw(Field):
+    """:class:`Field` raw input data"""
+
+    def __init__(self, *args, **kwargs):
+        super(Raw, self).__init__(*args, **kwargs)
 
 
 class List(Field):
@@ -197,12 +261,16 @@ class List(Field):
     """
 
     def __init__(self, *args, **kwargs):
-        kwargs.setdefault('default', list)
+        kwargs.setdefault('default', [])
         kwargs.setdefault('encoders', []).append(builtin_encoders.List())
 
         super(List, self).__init__(
             builtin_validators.List(*kwargs.get('inner_validators', [])),
             *args, **kwargs)
+    
+    @property
+    def field_type(self):
+        return list
 
 
 class Collection(Field):
@@ -237,7 +305,7 @@ class Collection(Field):
     """
 
     def __init__(self, model, *args, **kwargs):
-        kwargs.setdefault('default', list)
+        kwargs.setdefault('default', [])
 
         kwargs.setdefault('encoders', []).append(builtin_encoders.Collection())
         kwargs.setdefault('decoders', []).append(builtin_decoders.Collection(model))
@@ -257,3 +325,6 @@ class Collection(Field):
                 item = self.model(**item)
             result.append(item)
         return result
+
+__all__ = ("Field", "String", "Integer", "Float", "Boolean", "Embedded", "Email", "URL", "List", "Collection", "IP",
+           "Raw", "URI")